requirements.txt
pydantic==2.9.1
qdrant-client==1.11.1
pypdfium2==4.30.0
fastapi[standard]==0.114.0
python-dotenv==1.0.1
colpali_engine==0.3.0
openai==1.44.1
opencv_python_headless==4.10.0.84
tqdm==4.66.5
transformers==4.44.2
torch